+++ test/connection.test.js
@@ -153,7 +153,7 @@ describe('connections:', function() {
           }).
           then(function() {
             return new Promise(function(resolve) {
-              setTimeout(function() { resolve(); }, 4000);
+              setTimeout(function() { resolve(); }, 8000);
             });
           }).
           then(function() {
@@ -356,7 +356,7 @@ describe('connections:', function() {
   });
 
   it('should accept mongodb://aaron:psw@localhost:27000/fake', function(done) {
-    var db = mongoose.createConnection('mongodb://aaron:psw@localhost:27000/fake', { useNewUrlParser: true }, () => {
+    const db = mongoose.createConnection('mongodb://aaron:psw@localhost:27000/fake', { useNewUrlParser: true }, () => {
       db.close(done);
     });
     assert.equal(db.pass, 'psw');